ajuste al modelo de envia.com

+++ ServicesManagementCentral.Web/Models/Cotizador/CostSummaryModel.cs
@@ -9,18 +9,18 @@ namespace ServicesManagement.Web.Models.Cotizador
     {
         public int quantity { get; set; }
         public decimal basedPrice { get; set; }
-        public int extendedFare { get; set; }
-        public int insurance { get; set; }
-        public int additionalServices { get; set; }
+        public decimal extendedFare { get; set; }
+        public decimal insurance { get; set; }
+        public decimal additionalServices { get; set; }
         public decimal totalPrice { get; set; }
         public string currency { get; set; }
         public bool customKey { get; set; }
-        public int cashOnDeliveryCommission { get; set; }
-        public int cashOnDeliveryAmount { get; set; }
-        public int smsCommission { get; set; }
-        public int importFree { get; set; }
-        public int taxes { get; set; }
-        public int whatsappCommission { get; set; }
+        public decimal cashOnDeliveryCommission { get; set; }
+        public decimal cashOnDeliveryAmount { get; set; }
+        public decimal smsCommission { get; set; }
+        public decimal importFree { get; set; }
+        public decimal taxes { get; set; }
+        public decimal whatsappCommission { get; set; }
         public string folio { get; set; }
     }
 }
